moved new package to existing one
