+++ spring-boot-rest/src/main/java/com/baeldung/web/config/MyCustomErrorAttributes.java
@@ -1,4 +1,4 @@
-package com.baeldung.errorhandling.boot.configurations;
+package com.baeldung.web.config;
 
 import java.util.Map;
