+++ spring-boot-rest/src/main/resources/application.properties
@@ -1,0 +1,3 @@
+### Spring Boot default error handling configurations
+#server.error.whitelabel.enabled=false
+#server.error.include-stacktrace=always
