--- spring-boot-rest/src/main/resources/errorhandling-application.properties
@@ -1,3 +1,0 @@
-
-#server.error.whitelabel.enabled=false
-#server.error.include-stacktrace=always
