+++ spring-boot-rest/src/test/java/com/baeldung/web/error/ErrorHandlingLiveTest.java
@@ -1,4 +1,4 @@
-package com.baeldung.errorhandling.boot;
+package com.baeldung.web.error;
 
 import static io.restassured.RestAssured.given;
 import static org.assertj.core.api.Assertions.assertThat;
